+++ CHANGELOG.md
@@ -30,6 +30,19 @@ Then in both cases it will use `7.0.0` when installing the `sentry-android-okhtt
 - Add aarch64 sentry-cli ([#611](https://github.com/getsentry/sentry-android-gradle-plugin/pull/611))
     - This is used when the build is executed inside a docker container on an Apple silicon chip (e.g. M1)
 
+- Instrument ContentProvider/Application onCreate calls to measure app-start performance ([#565](https://github.com/getsentry/sentry-android-gradle-plugin/pull/565))
+    - This feature requires the `sentry-java` SDK version `7.1.0` and is enabled by default
+    - To disable the feature, set `sentry.tracingInstrumentation.appStart.enabled` to `false`
+```kotlin
+sentry {
+  tracingInstrumentation {
+    appStart {
+      enabled.set(false)
+    }
+  }
+}
+```
+
 ### Fixes
 
 - Fix sentry-cli url parameter position ([#610](https://github.com/getsentry/sentry-android-gradle-plugin/pull/610))